Proper price threshold update
Doing it in onPause was bad, as the ViewModel paused as well and the update
never happened. Use a dedicated TextWatcher and only programmatically set
when there is no focus (i.e. when the user didn't trigger a text change).

Lastly, change the icon if the user set an alarm to give a visual clue.